Determine if a given grade is an 'A'
This makes the GradeBookTests.OverrideRankedGradeBooksGetLetterGrade-
Tests.TopTwentyPercentGetAnATests test pass.

We follow the general outline given in the course by first sorting the
students by grade.  We then find the index of the first student whose
grade is at most the given grade.  If that index is within the first 20%
of indexes, then the given grade is an 'A'.

Also: Add 'using System' directive.